requirements.txt
pytz==2023.3
Pillow==9.5.0
zipp==3.16.2
matplotlib==3.7.2
pyparsing==3.0.9
Babel==2.12.1
python-dateutil==2.8.2
numpy==1.24.4
six==1.16.0
tkcalendar==1.6.1
kiwisolver==1.4.4
contourpy==1.0.7
wheel==0.38.4
cycler==0.11.0
pygame==2.5.1
pydub==0.25.1
mutagen==1.47.0